hf-spaces/brain-tumor-model/requirements.txt
gradio
tensorflow
pillow
numpy 